+++ lib/src/models/context_menu_item.dart
@@ -1,0 +1,17 @@
+import 'package:equatable/equatable.dart';
+
+class FormattedTextContextMenuItem extends Equatable {
+  const FormattedTextContextMenuItem({
+    required this.label,
+    required this.patternChars,
+  });
+
+  final String label;
+  final String patternChars;
+
+  @override
+  List<Object?> get props => [
+        label,
+        patternChars,
+      ];
+}